refactor: Add support for detail images in CreateEventDto
This commit adds support for detail images in the CreateEventDto. The CreateEventDto now includes a new property called detailImages, which is an array of objects representing the detail images of the event. Each object contains the imageUrl and order properties. This change required modifications to the CreateEventDto class and the Event entity.

The axios package has also been added as a new dependency in the package.json file.